Remove parameters from Simulation panel and put them in LayoutContext
- Also move potentially useful components elsewhere:
  - SliderControl (into a branch on UI)
  - useDebouncedEffect into hooks/ folder
