+++ src/components/RealtimeCanvas/SimulationPanel.js
@@ -1,23 +1,18 @@
 /* eslint-disable no-param-reassign */
 import React, {
-  useRef,
   useState,
   useContext,
   useCallback,
-  useReducer,
-  useEffect,
 } from 'react';
 import PropTypes from 'prop-types';
 import cx from 'classnames';
 import { motion, useDragControls } from 'framer-motion';
-import { debounce } from 'lodash';
 import Toggle from '@codaco/ui/lib/components/Fields/Toggle';
 import AutorenewIcon from '@material-ui/icons/AutorenewRounded';
 import ZoomInIcon from '@material-ui/icons/ZoomInRounded';
 import ZoomOutIcon from '@material-ui/icons/ZoomOutRounded';
 import MinimizeIcon from '@material-ui/icons/MinimizeRounded';
 import LayoutContext from '../../contexts/LayoutContext';
-import Slider from '../SliderControl';
 
 const panelVariants = {
   minimized: {
@@ -80,86 +75,6 @@ ButtonControl.defaultProps = {
   disabled: false,
 };
 
-const SliderControl = ({
-  value,
-  label,
-  min,
-  max,
-  name,
-  disabled,
-  onChange,
-}) => {
-  const handleChange = useCallback((nextValue) => {
-    onChange(nextValue, name);
-  }, [name]);
-
-  return (
-    <div
-      className={cx(
-        'simulation-panel__control',
-        'simulation-panel__control--slider',
-        { 'simulation-panel__control--disabled': disabled },
-      )}
-      onPointerDown={(e) => { e.stopPropagation(); }}
-    >
-      {label}
-      <Slider
-        value={value}
-        min={min}
-        max={max}
-        onChangeEnd={handleChange}
-      />
-    </div>
-  );
-};
-
-SliderControl.propTypes = {
-  min: PropTypes.number,
-  max: PropTypes.number,
-  value: PropTypes.number.isRequired,
-  label: PropTypes.string.isRequired,
-  disabled: PropTypes.bool,
-};
-
-SliderControl.defaultProps = {
-  min: 0,
-  max: 100,
-  disabled: false,
-};
-
-const optionsReducer = (state, { type, payload }) => {
-  switch (type) {
-    case 'UPDATE':
-      return {
-        ...state,
-        [payload.option]: payload.value,
-      };
-    default:
-      return state;
-  }
-};
-
-const useDebouncedEffect = (func, wait, deps) => {
-  const currentFunc = useRef(null);
-  const debouncedFunc = useRef(debounce(() => { currentFunc.current(); }, wait));
-  useEffect(() => {
-    if (!currentFunc.current) { // skip first run
-      currentFunc.current = func;
-      return;
-    }
-    currentFunc.current = func;
-    debouncedFunc.current();
-  }, deps);
-};
-
-const initialOptions = {
-  decay: 0.1,
-  charge: -30,
-  linkForce: 1,
-  linkDistance: 30,
-  center: 0.1,
-};
-
 const SimulationPanel = ({
   dragConstraints,
 }) => {
@@ -172,28 +87,11 @@ const SimulationPanel = ({
   const [isMinimized, setMinimized] = useState(false);
   const dragControls = useDragControls();
 
-  const [options, optionsAction] = useReducer(optionsReducer, initialOptions);
-
   const toggleMinimize = () => setMinimized((minimized) => !minimized);
 
   const handleStartDrag = useCallback((event) => {
     dragControls.start(event);
   }, [dragControls]);
-
-  const handleChangeSlider = useCallback((value, option) => {
-    optionsAction({
-      type: 'UPDATE',
-      payload: {
-        option,
-        value,
-      },
-    });
-  }, [optionsAction]);
-
-  useDebouncedEffect(() => {
-    if (!simulation) { return; }
-    simulation.updateOptions(options);
-  }, 50, [options]);
 
   if (!allowAutomaticLayout) { return null; }
 
@@ -271,38 +169,6 @@ const SimulationPanel = ({
             onClick={reheat}
             disabled={isRunning}
             label="Reheat"
-          />
-          <SliderControl
-            label="Velocity Decay"
-            name="decay"
-            value={options.decay}
-            onChange={handleChangeSlider}
-            min={0}
-            max={1}
-          />
-          <SliderControl
-            label="Charge Strength"
-            name="charge"
-            value={options.charge}
-            onChange={handleChangeSlider}
-            min={-50}
-            max={0}
-          />
-          <SliderControl
-            label="Link Distance"
-            name="linkDistance"
-            value={options.linkDistance}
-            onChange={handleChangeSlider}
-            min={0}
-            max={30}
-          />
-          <SliderControl
-            label="Center Strength"
-            name="center"
-            value={options.center}
-            onChange={handleChangeSlider}
-            min={0}
-            max={1}
           />
         </motion.div>
       </motion.div>
