+++ src/hooks/useForceSimulation.js
@@ -31,7 +31,7 @@ const useForceSimulation = (listener = () => {}) => {
 
   useEffect(() => viewport.zoom.onChange(() => recalculate()), [recalculate]);
 
-  const initialize = useCallback((network, options = {}) => {
+  const initialize = useCallback((network = {}, options = {}) => {
     if (worker.current) { worker.current.terminate(); }
 
     worker.current = new ForceSimulationWorker();
