+++ src/styles/components/_all.scss
@@ -37,7 +37,6 @@
 @import 'case-id-form';
 @import 'stack-button';
 @import 'simulation-panel';
-@import 'slider-control';
 
 // Session Panel
 @import './session-panel/all';
